Implement trivial delta bundle support (Fixes #334)
This doesn't get any actual benefits of using delta bundles, it simply
repackages the current bundle in a trivial delta bundle. This seems to
have worse performance that using the standard format, so I added a
one-time warning (bike-shedding welcome on the wording)

This is still helpful in the case of automated tests which (re)install
and run the application immediately without user interaction, where
there is no chance to disable delta bundles.

If it helps, however, the format (as reverse-engineered from the
android client code) is described in case full delta support is added
later, or someone wants to know what the heck this is trying to do :)

Note that I also made the adb reverse call one-time, since it's pretty
spammy when reloading a lot!