fix makeHot to not re-render unnecessarily (#491)
This will fix the `makeHot` so it doesn't force-rerender every tiny little change that happens when used with routers.  It does this by not building a brand new component on every call of the resulting function. 

```javascript
const HotView = makeHot(() => HomeView, 'Home');
console.log(HotView() === HotView());
```

> Before the above would return false, therefore a re-render would be forced every time . `getScreen` is called by `react-navigation`

Note that this will still cause problems when utilized as documented since

```javascript
console.log(makeHot(() => HomeView, 'Home')() !== makeHot(() => HomeView, 'Home')())
```

Solution there would likely to use a `Map` of some sort to identify when we should return the previously created component instead of creating a new one.  Since I couldn't be 100% of the effects of doing this type of change I decided to at least make it fixable in the meantime. 

See #490