Add transpilation of @react-native-community
The new @react-native-community slider & asyncstorage libs for RN59 are shipped with flow types and require transpilation. This PR makes sure that all those libs are transpiled by default. This should typically not lead to any performance issues, as this code is almost always a small javascript interface to the native code.