fix: Avoid tripping Xcode's script error detection. (#525)
Xcode seems to fail if the verbosity level is set to anything other than `verbose`. Some of the characters used in the other levels makes Xcode think that errors have occurred. This was filed as #514 but was closed a bit early since there is a workaround.

### Summary

The patch detects whether we're in interactive mode and applies the correct verbosity level.

### Test plan

By default, without adding `--progress verbose` to the bundle command, Xcode should not fail.