docs: add note about localhost/0.0.0.0 issue (#622)
This commit updates the Configuration doc to include a note about
configuring the server to use host '0.0.0.0' if there is an issue
connecting to the dev server on localhost.

I experienced an issue where my react native app running on a ios device
could not connect to the development server when using the default
(localhost) server configuration. Changing the host to 0.0.0.0 fixed the
issue for me so others might find that advice useful as well.