chore: remove haul inspector (#721)
The reasoning behind removing haul inspector is that, it doesn't give much benefit over doing standard Node debugging and/or outputting logs to a file. In near future we plan to improve logging, including more debug logs, which will be more verbose, so that the users can run haul and investigate those verbose logs in a file.