feat: Allow browser to be specified
Instead of making the assumption on behalf of the user, allow them to
control the browser which is used for the debugger UI. This is
controlled via the `RCT_DEBUGGER_BROWSER` environment variable. If this
variable is populated, then it will be tried first, before falling back
to the baked-in defaults.